+++ src/Symfony/Component/PropertyAccess/Tests/DependencyInjection/Fixtures/property_accessor.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('property_access', null);